src/main/webapp/app/entities/sante/list/sante.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { SanteService } from '../service/sante.service';

import { SanteComponent } from './sante.component';

describe('Sante Management Component', () => {
  let comp: SanteComponent;
  let fixture: ComponentFixture<SanteComponent>;
  let service: SanteService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [RouterTestingModule.withRoutes([{ path: 'sante', component: SanteComponent }]), HttpClientTestingModule],
      declarations: [SanteComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            data: of({
              defaultSort: 'id,asc',
            }),
            queryParamMap: of(
              jest.requireActual('@angular/router').convertToParamMap({
                page: '1',
                size: '1',
                sort: 'id,desc',
              })
            ),
            snapshot: { queryParams: {} },
          },
        },
      ],
    })
      .overrideTemplate(SanteComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(SanteComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(SanteService);

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        })
      )
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.santes?.[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  describe('trackId', () => {
    it('Should forward to santeService', () => {
      const entity = { id: 123 };
      jest.spyOn(service, 'getSanteIdentifier');
      const id = comp.trackId(0, entity);
      expect(service.getSanteIdentifier).toHaveBeenCalledWith(entity);
      expect(id).toBe(entity.id);
    });
  });
});
